requirements.txt
# Deep Learning Framework
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Web Interface
streamlit>=1.28.0

# Utilities
PyYAML>=6.0
tqdm>=4.65.0
scikit-learn>=1.3.0

# Optional: For better performance
# torch-tensorrt  # NVIDIA TensorRT for faster inference
# onnx>=1.14.0    # ONNX export support
# onnxruntime     # ONNX runtime
